split dynamic from start
