+++ huginn-proxy-lib/src/config/mod.rs
@@ -1,6 +1,7 @@
 pub mod dynamic;
 mod effective;
 pub mod parser;
+mod secret;
 pub mod startup;
 pub mod watcher;
 
@@ -20,6 +21,7 @@ pub use effective::{EffectiveConfigSummary, EffectiveConfigView};
 pub use loader::{load_from_path, security_override_warnings, SecurityOverrideWarning};
 pub use parser::{ConfigFormat, ConfigParser, TomlParser, YamlParser};
 pub use root::{Config, ConfigParts};
+pub use secret::Secret;
 pub use startup::{
     ClientAuth, FingerprintConfig, KeepAliveConfig, ListenConfig, LoggingConfig,
     ProxyProtocolConfig, ProxyProtocolMode, SessionResumptionConfig, StaticConfig, TelemetryConfig,
