+++ huginn-proxy-lib/tests/config/mod.rs
@@ -2,4 +2,5 @@ mod effective;
 mod header_manipulation;
 mod loader;
 mod parser;
+mod secret;
 mod types;
